requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tldextract>=5.0.0
pypdf>=3.17.0
flask==2.3.2
firebase-admin==6.3.0
python-docx
openpyxl
tenacity
textract

openai
beautifulsoup4
playwright
langdetect
textblob
# python -m playwright install 